vertex_ai_agent/pyproject.toml
[project]
name = "pi-system-vertex-ai-agent"
version = "1.0.0"
description = "Vertex AI Agent for AVEVA PI System with MCP integration"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
requires-python = ">=3.10"
dependencies = [
    "google-genai[adk]>=0.2.0",
    "mcp>=0.1.0",
    "aiohttp>=3.9.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "google-cloud-aiplatform>=1.38.0",
    "google-cloud-storage>=2.10.0",
    "google-cloud-logging>=3.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"
